Starting implementing loot hunting - WIP
The plan: If some loot item is required for doing a certain mission,
then the script will automatically start trying to get that loot item
by doing a mission that gives it as a reward. As soon as the item is
received, auto mission continues doing missions in normal order.